# backend-go/src/docs/swagger.yaml
basePath: /api
definitions:
...
  routes.DeleteReportResponse:
    properties:
      message:
        description: 消息
        type: string
    type: object
...
  routes.ErrorResponse:
    properties:
      error:
        description: 错误信息
        type: string
    type: object
  routes.ExportReportRequest:
    properties:
      format:
        description: 格式
        type: string
      id:
        description: 报告ID
        type: string
      token:
        description: Token
        type: string
    type: object
  routes.ExportReportResponse:
    properties:
      content:
        description: 内容
        type: string
      filename:
        description: 文件名
        type: string
      message:
        description: 消息
        type: string
    type: object
...
  routes.GenerateReportRequest:
    properties:
      end_date:
        description: 结束日期
        type: string
      period:
        description: 周期
        type: string
      start_date:
        description: 开始日期
        type: string
      token:
        description: Token
        type: string
      type:
        description: 报告类型
        type: integer
    type: object
  routes.GenerateReportResponse:
    properties:
      message:
        description: 消息
        type: string
      report:
        allOf:
        - $ref: '#/definitions/routes.Report'
        description: 报告信息
    type: object
  routes.GetReportResponse:
    properties:
      message:
        description: 消息
        type: string
      report:
        allOf:
        - $ref: '#/definitions/routes.Report'
        description: 报告信息
    type: object
  routes.GetReportsResponse:
    properties:
      message:
        description: 消息
        type: string
      reports:
        description: 报告列表
        items:
          $ref: '#/definitions/routes.Report'
        type: array
    type: object
...
  routes.PolishReportRequest:
    properties:
      api_key:
        description: API密钥
        type: string
      api_url:
        description: API URL
        type: string
      model:
        description: 模型
        type: string
      provider:
        description: 提供商
        type: string
      report_id:
        description: 报告ID
        type: string
      token:
        description: Token
        type: string
    type: object
  routes.PolishReportResponse:
    properties:
      message:
        description: 消息
        type: string
      report:
        allOf:
        - $ref: '#/definitions/routes.Report'
        description: 报告信息
    type: object
...
  routes.Report:
    properties:
      content:
        description: 内容
        type: string
      id:
        description: 报告ID
        type: string
      period:
        description: 周期
        type: string
      polished_content:
        description: 润色后的内容
        type: string
      statistics:
        allOf:
        - $ref: '#/definitions/routes.ReportStatistics'
        description: 统计信息
      tasks:
        description: 任务列表
        items:
          type: string
        type: array
      title:
        description: 标题
        type: string
      type:
        description: 报告类型
        type: integer
      user_id:
        description: 用户ID
        type: string
    type: object
  routes.ReportStatistics:
    properties:
      completed_tasks:
        description: 完成任务数
        type: integer
      completion_rate:
        description: 完成率
        type: integer
      in_progress_tasks:
        description: 进行中任务数
        type: integer
      overdue_tasks:
        description: 逾期任务数
        type: integer
      total_tasks:
        description: 总任务数
        type: integer
    type: object
...
host: localhost:5001
info:
  contact: {}
  description: TodoIng task management system API
  title: TodoIng API
  version: "1.0"
paths:
...
  /reports:
    get:
      consumes:
      - application/json
      description: 获取用户的所有报告
      parameters:
      - description: 用户ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.GetReportsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 获取报告列表
      tags:
      - reports
  /reports/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除报告
      parameters:
      - description: 报告ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.DeleteReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 删除报告
      tags:
      - reports
    get:
      consumes:
      - application/json
      description: 根据ID获取报告详情
      parameters:
      - description: 报告ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.GetReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 获取单个报告
      tags:
      - reports
  /reports/export:
    post:
      consumes:
      - application/json
      description: 导出报告为指定格式
      parameters:
      - description: 导出报告请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.ExportReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ExportReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 导出报告
      tags:
      - reports
  /reports/generate:
    post:
      consumes:
      - application/json
      description: 根据类型和周期生成报告
      parameters:
      - description: 生成报告请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.GenerateReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.GenerateReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 生成报告
      tags:
      - reports
  /reports/polish:
    post:
      consumes:
      - application/json
      description: 使用AI润色报告内容
      parameters:
      - description: 润色报告请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.PolishReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.PolishReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: 润色报告
      tags:
      - reports
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"